Refactor IA service structure and improve validation
