+++ ia-service/tests/test_no_api_key.py
@@ -19,12 +19,14 @@ class DummyAsyncClient:
 
 sys.modules['httpx'] = types.SimpleNamespace(AsyncClient=DummyAsyncClient)
 
-from app.main import (
+from app.routers.ia import (
     analyze_performance,
     predict_match,
     detect_errors,
     suggest_tactics,
     suggest_lineup,
+)
+from app.schemas import (
     PerformanceRequest,
     MatchPredictionRequest,
     ErrorDetectionRequest,
@@ -34,35 +36,35 @@ from app.main import (
 
 
 def test_analyze_performance_no_key(monkeypatch):
-    monkeypatch.setattr('app.main.OPENAI_API_KEY', None)
+    monkeypatch.setattr('app.routers.ia.OPENAI_API_KEY', None)
     payload = PerformanceRequest(ratings=[{"player": "John", "score": 7}])
     with pytest.raises(HTTPException) as exc:
         asyncio.run(analyze_performance(payload, DummyAsyncClient()))
     assert exc.value.status_code == 500
 
 def test_predict_match_no_key(monkeypatch):
-    monkeypatch.setattr('app.main.OPENAI_API_KEY', None)
+    monkeypatch.setattr('app.routers.ia.OPENAI_API_KEY', None)
     payload = MatchPredictionRequest(home_team=['A'], away_team=['B'])
     with pytest.raises(HTTPException) as exc:
         asyncio.run(predict_match(payload, DummyAsyncClient()))
     assert exc.value.status_code == 500
 
 def test_detect_errors_no_key(monkeypatch):
-    monkeypatch.setattr('app.main.OPENAI_API_KEY', None)
+    monkeypatch.setattr('app.routers.ia.OPENAI_API_KEY', None)
     payload = ErrorDetectionRequest(lineup=['A', 'B'], formation=None)
     with pytest.raises(HTTPException) as exc:
         asyncio.run(detect_errors(payload, DummyAsyncClient()))
     assert exc.value.status_code == 500
 
 def test_suggest_lineup_no_key(monkeypatch):
-    monkeypatch.setattr('app.main.OPENAI_API_KEY', None)
+    monkeypatch.setattr('app.routers.ia.OPENAI_API_KEY', None)
     payload = LineupRequest(players=['A', 'B'], formation='4-4-2')
     with pytest.raises(HTTPException) as exc:
         asyncio.run(suggest_lineup(payload, DummyAsyncClient()))
     assert exc.value.status_code == 500
 
 def test_suggest_tactics_no_key(monkeypatch):
-    monkeypatch.setattr('app.main.OPENAI_API_KEY', None)
+    monkeypatch.setattr('app.routers.ia.OPENAI_API_KEY', None)
     payload = TacticsRequest(players=['John'], style=None)
     with pytest.raises(HTTPException) as exc:
         asyncio.run(suggest_tactics(payload, DummyAsyncClient()))
